+++ q2doc/usage/templates/init_metadata.html
